Allow overriding dynamic actor shapes when loading checkpoint
The shape of the tensor depends on exact time ranges for training,
as well as some other dataparser settings. It makes more sense to
load the poses that were actually used for training the nerf, and
only warn if the shapes differ.

Current behaviour is to crash, which can easily make old checkpoints
unusable.